Adds the Quake Parser class
This class is the core of this project, with it we could retrieve
information from Quake Server log files. It uses the previous builded
parsers and entities to handle events fetched from lines from a log
file.

It implements some logs to avoid player name duplications. Also, it
tries to detect when a log has games without properly ending lines,
and register kills only by the following logic:

1) The <world> kill isn't showed in the kills hash.
2) If the player kill itselfs the kill won't count for it
(maybe should be -1?).
3) If the <world> kills a player, the player score will be decreased
by 1.